Bug Fix - Shift+3 causes odd behavior when using Libki client
Once the customer is logged into Libki and they open browser,
document, whatever and they hit SHIFT+3, the keyboard locks up and thinks
that the control key is held down.  All keystrokes after this try to
execute the keyboard shortcut (i.e. hitting the "H" key will act as
CONTROL+H and open the history in Chrome).  The only fix we have found is
to hold down the left and right control keys to kind of clear this.  But
as soon as you hit SHIFT+3 again, the problem returns.