Bug fix - Disable messaging feature, make client more resiliant to malformed messages
Sometimes client pops of a "parse error" message. This is believed to be
an issue with the messaging system, which is not fully implemented and
thus never used. Commenting out the code that parses and displays
messages.

Note: Code for messaging system will also need to be removed from the
server at some point.