Add ability to restrict client usage by age
To use this feature, just add a new key under the 'node' section of the
Libki client ini named 'age_limit'.

Example:
[node]
age_limit="gt18"

This will limit the client to patrons older than 18. At the moment this
only works via SIP2 as there is no way to edit a user's age from the staff
administration ( yet ).

Multiple age limits can be implemented delimited by commans, such as:
age_limit="ge11,le17" which will limit the client to users between ( and
including ) the ages of 11 and 17.

The format first two characters are the comparison. Supported comparisons
are:
eq: equal to
ne: not equal to
lt: less than
gt: greater than
le: less than or equal to
ge: greater than or equal to

It is possible to make a client unusable by anyone ( e.g. "gt18,lt17" ) so
be careful with this configuration.