Add the ability for the client to disable input for a given number of seconds on startup
Some computers can be too slow to load the Libki client. In these cases,
it's possible for a user to launch a program before the client has
launched. Then, the program sites over top of the client login screen thus
allowing the user to circumvent the client.

This feature adds the ability to specify DisableInput <seconds> in the
[windows] section of the client ini file.

For example, to disable input for 10 seconds:
[windows]
DisableInput=10