Add appveyor.yml
Make appveyor build the installer

* Enable appveyor to deploy each client installer build
* Use env variable instead for github token
* Make AppVeyor skip tags to prevent infinite loop

From the AppVeyor documentation:
---
NOTE: If the current build is based on a regular commit then a new tag is created (either explicitly or automatically named in tag field of deployment settings as described below) along with the release. This means that, unless you have skip_tags set to true, the creation of the tag will also send a webhook to Appveyor, and the deployment process will begin again, and so on.

To avoid this possible cycle, you can set skip_tags to true. However, you may be using on: key to conditionally build and deploy only on tags you create manually. In this case, you can specify tag: $(APPVEYOR_REPO_TAG_NAME) in deployment settings without skipping tags, and the manually created tag will be updated instead of a new one being created.
---