Add ability to store currently logged in user in ini
This commit causes the Libki client to store the currently logged in
user in app written ini data. On windows this file is usually located at
%APPDATA%/Libki/Libki Kiosk Management System.ini

with the "session" block will be a key "LoggedInUser", e.g. :
--
[session]
LoggedInUser=kylemhall
--

This feature will allow other software to know who is currently logged
into the Libki client.