Add ability to specify and inactivity timeout for the Libki client
This feature adds two new optional keys under the 'node' section of the
ini file: inactivity_logout and inactivity_timeout.

The first is the max minutes of inactivity until the client is logged out
automatically. The second is the number of minutes of inactivity before
the user is warned they will be logged out due to inactivity.