Moved ToS text to Detailed Text field
I had an usual problem where a really long ToS (as the one we're required to use) caused the box to overflow, and the yes and no buttons fell off the bottom of the screen, making them impossible to click. By moving them into the Detailed text field, we get that scrollbox instead, which prevents the buttons from falling off the bottom of the screen.