Add ability for a user to temporarily lock session
This feature adds the ability for a user to remotely lock his or her
current session in order to leave the computer and return. Useful for
retrieving prints, using facilities, etc. The users time continues to
count down while the client is locked.

This feature is off by default and is enabled from a system setting
on the Libki server.